prevent redis test from appearing when unavailable
The Redis object cache test previously displayed a "good" status on plans where Redis is not available (e.g. Basic plans). This was confusing as it reported on a feature that could not be used.

This change modifies the test to return early and not register itself in this scenario. This prevents the test from appearing in the Site Health screen, providing a cleaner and more accurate user experience.