Improve Site Health check for plans without Redis (#96)
* improve object cache test for plans without Redis

Previously, the Site Health check for the Redis Object Cache would show a
'critical' error if the `CACHE_HOST` was not defined. This was confusing
for users on Basic plans where Redis is intentionally unavailable.

This change checks the service level. If the plan is one where Redis is
not available (e.g., 'basic'), the Site Health tool now displays a
'good' status with an informational message. This message explains that
Redis is not included in their plan and suggests upgrading.

This provides a better user experience and avoids unnecessary critical
warnings for users on these plans.

* add test for object cache on plans without Redis

Adds a new test case to verify that the object cache health check correctly reports a 'good' status for plans (e.g., Basic) that do not include Redis.

Existing tests are updated to explicitly run under a 'performance' plan context to ensure they test for Redis availability as intended, isolating the new test case.

* refine object cache test for plans without redis

Use `unset()` instead of setting the environment variable to `null` to
more accurately reflect a missing `CACHE_HOST` variable.

The `assertContains()` method is less brittle when messages change

* fix the stringcontainsstring test

* fix linting

* upgrade Pantheon WP Coding Standards
and fix linting

* use pantheon wp coding standards 2 for php 7.4

* add helper function to get request headers

Introduces the `_pantheon_get_request_headers` function to retrieve and normalize HTTP request headers from the `$_SERVER` superglobal.

This utility standardizes header keys from the `HTTP_*` format (e.g., `HTTP_USER_AGENT`) to the standard HTTP header format (e.g., `User-Agent`), making them easier to work with in subsequent code.

* add helper function to get a specific header

This commit introduces the `_pantheon_get_header` function to provide a convenient way to access a single request header value.

* prevent redis test from appearing when unavailable

The Redis object cache test previously displayed a "good" status on plans where Redis is not available (e.g. Basic plans). This was confusing as it reported on a feature that could not be used.

This change modifies the test to return early and not register itself in this scenario. This prevents the test from appearing in the Site Health screen, providing a cleaner and more accurate user experience.

* the label for all pantheon tests should be pantheon

* change ENV to SERVER

* update tests to check for empty test result

we no longer report good or bad if we're on basic or headers are missing entirely

* Apply suggestions from code review

Co-authored-by: Phil Tyler <phil.tyler@pantheon.io>

* fix line spacing

* add some inline commenting

---------

Co-authored-by: Phil Tyler <phil.tyler@pantheon.io>